(some) sanity to nav and bring login screen for the unauthenticated
