+++ src/app/(app)/admin/_layout.tsx
@@ -3,7 +3,12 @@ import { ProtectedRoute } from "../../../hooks/auth/protectedRoute";
 
 export default function AdminLayout() {
   return (
-    <ProtectedRoute backendName="defaultBackend" requireAuthType="mtls" requireValidUser={true} allowedUserTypes={["admin"]}>
+    <ProtectedRoute
+      backendName="defaultBackend"
+      requireAuthType="mtls"
+      requireValidUser={true}
+      allowedUserTypes={["admin"]}
+    >
       <Slot />
     </ProtectedRoute>
   );
