+++ src/app/(app)/users/_layout.tsx
@@ -1,10 +1,53 @@
+import { Tabs } from "expo-router";
+import React from "react";
+
 import { Slot } from "expo-router";
 import { ProtectedRoute } from "../../../hooks/auth/protectedRoute";
+import { TabBarIcon } from "@/src/components/navigation/TabBarIcon";
+import { Colors } from "@/src/constants/Colors";
+import { useColorScheme } from "@/src/hooks/useColorScheme";
 
-export default function UsersLayout() {
+export default function TabLayout() {
+  const colorScheme = useColorScheme();
+
   return (
-    <ProtectedRoute backendName="defaultBackend" requireAuthType="mtls" requireValidUser={true} allowedUserTypes={["user"]}>
-      <Slot />
+    <ProtectedRoute
+      backendName="defaultBackend"
+      requireAuthType="mtls"
+      requireValidUser={true}
+      allowedUserTypes={["user"]}
+    >
+      <Tabs
+        screenOptions={{
+          tabBarActiveTintColor: Colors[colorScheme ?? "light"].tint,
+          headerShown: false,
+        }}
+      >
+        <Tabs.Screen
+          name="index"
+          options={{
+            title: "Home",
+            tabBarIcon: ({ color, focused }) => (
+              <TabBarIcon
+                name={focused ? "home" : "home-outline"}
+                color={color}
+              />
+            ),
+          }}
+        />
+        <Tabs.Screen
+          name="explore"
+          options={{
+            title: "Explore",
+            tabBarIcon: ({ color, focused }) => (
+              <TabBarIcon
+                name={focused ? "code-slash" : "code-slash-outline"}
+                color={color}
+              />
+            ),
+          }}
+        />
+      </Tabs>
     </ProtectedRoute>
   );
 }
