+++ src/app/(app)/users/index.tsx
@@ -1,10 +1,76 @@
-import React from "react";
-import { View, Text } from "react-native";
+import { Image, StyleSheet, Platform } from "react-native";
 
-export default function UsersIndex() {
+import { HelloWave } from "@/src/components/HelloWave";
+import ParallaxScrollView from "@/src/components/ParallaxScrollView";
+import { ThemedText } from "@/src/components/ThemedText";
+import { ThemedView } from "@/src/components/ThemedView";
+import { Link } from "expo-router";
+
+export default function HomeScreen() {
   return (
-    <View>
-      <Text>User Home</Text>
-    </View>
+    <ParallaxScrollView
+      headerBackgroundColor={{ light: "#A1CEDC", dark: "#2b2e31" }}
+      headerImage={
+        <Image
+          source={require("@/src/assets/mainlogo_nob.png")}
+          style={styles.reactLogo}
+        />
+      }
+    >
+      <ThemedView style={styles.titleContainer}>
+        <ThemedText type="title">Deploy App II</ThemedText>
+      </ThemedView>
+      <ThemedView style={styles.stepContainer}>
+        <ThemedText type="subtitle">Step 1: Try it</ThemedText>
+        <ThemedText>
+          Edit{" "}
+          <ThemedText type="defaultSemiBold">app/(tabs)/index.tsx</ThemedText>{" "}
+          to see changes. Press{" "}
+          <ThemedText type="defaultSemiBold">
+            {Platform.select({ ios: "cmd + d", android: "cmd + m" })}
+          </ThemedText>{" "}
+          to open developer tools.
+        </ThemedText>
+      </ThemedView>
+      <ThemedView style={styles.stepContainer}>
+        <ThemedText type="subtitle">Step 2: Explore</ThemedText>
+        <ThemedText>
+          Tap the Explore tkikkiab to learn more about what's included in this
+          starter app.
+        </ThemedText>
+      </ThemedView>
+      <ThemedView style={styles.stepContainer}>
+        <ThemedText type="subtitle">Step 3: Get a fresh start</ThemedText>
+        <ThemedText>
+          When you're ready, run{" "}
+          <ThemedText type="defaultSemiBold">npm run reset-project</ThemedText>{" "}
+          to get a fresh <ThemedText type="defaultSemiBold">app</ThemedText>{" "}
+          directory. This will move the current{" "}
+          <ThemedText type="defaultSemiBold">app</ThemedText> to{" "}
+          <ThemedText type="defaultSemiBold">app-example</ThemedText>.
+        </ThemedText>
+        <ThemedText type="subtitle">Step 3: Getdd a fresh start</ThemedText>
+        <Link href="/modal">Present modal</Link>
+      </ThemedView>
+    </ParallaxScrollView>
   );
 }
+
+const styles = StyleSheet.create({
+  reactLogo: {
+    bottom: 0,
+    height: 190,
+    left: 0,
+    position: "absolute",
+    width: 250,
+  },
+  stepContainer: {
+    gap: 8,
+    marginBottom: 8,
+  },
+  titleContainer: {
+    alignItems: "center",
+    flexDirection: "row",
+    gap: 8,
+  },
+});
